i1083: move sessionStorage handling to new sessionStorage util service
