+++ projects/WTI-UI/src/app/modules/options/components/options-page/options-page.component.ts
@@ -6,8 +6,8 @@ import { ChangePasswordComponent } from '../change-password/change-password.comp
 import { UiHelperService } from 'src/app/modules/core/services/ui-helper.service';
 import { environment } from 'src/environments/environment';
 import { AppTitleService } from 'src/app/modules/core/services/app-title.service';
-import { saveCurrentPage } from 'src/app/app.component';
-import { saveOptions } from 'src/app/app.component';
+import { saveCurrentPage } from 'src/app/modules/core/services/session-storage.utils';
+import { saveOptions } from 'src/app/modules/core/services/session-storage.utils';
 import * as Constants from 'src/constants';
 import { DEBUG_MODE } from 'src/constants';
