+++ projects/WTI-UI/src/app/modules/scoreboard/components/scoreboard-page/scoreboard-page.component.ts
@@ -3,7 +3,7 @@ import { IContestService } from 'src/app/modules/core/abstract-services/i-contes
 import { takeUntil } from 'rxjs/operators';
 import { Subject } from 'rxjs';
 import { AppTitleService } from 'src/app/modules/core/services/app-title.service';
-import { saveCurrentPage } from 'src/app/app.component';
+import { saveCurrentPage } from 'src/app/modules/core/services/session-storage.utils';
 import * as Constants from 'src/constants';
 import { DEBUG_MODE } from 'src/constants';
